[refactor] move abstract classes to directory
