+++ Backend/Entities/Abstract/AuditableEntity.cs
@@ -1,9 +1,8 @@
-
 using Backend.Interfaces.Entities;
 
-namespace Backend.Entities;
+namespace Backend.Entities.Abstract;
 
-public abstract class BaseAuditableEntity : BaseEntity, IAuditable
+public abstract class AuditableEntity : IAuditable
 {
   public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
   public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
