+++ Backend/Entities/BrewSetup.cs
@@ -1,3 +1,5 @@
+using Backend.Entities.Abstract;
+
 namespace Backend.Entities;
 
 public class BrewSetup : BaseEntity
