+++ Backend/Enums/EntityType.cs
@@ -1,0 +1,9 @@
+namespace Backend.Enums;
+
+public enum EntityType
+{
+  Brewer = 1,
+  Grinder = 2,
+  Bean = 3,
+  Roaster = 4,
+}
